+++ tests/logger.rs
@@ -2,6 +2,7 @@ use std::ffi::CStr;
 
 use libc::{c_char, c_void};
 use telio;
+use telio::TelioTracingSubscriber;
 
 mod test_module {
     use std::sync::atomic::{AtomicUsize, Ordering};
@@ -19,20 +20,18 @@ mod test_module {
             assert_eq!(telio::ffi_types::telio_log_level::TELIO_LOG_INFO, level);
             let message = CStr::from_ptr(message);
             assert_eq!(
-                r#""logger::test_module":37 test message"#,
+                r#""logger::test_module":36 test message"#,
                 message.to_str().unwrap()
             );
             let call_count: &AtomicUsize = &*(ctx as *const AtomicUsize);
             assert_eq!(0, call_count.fetch_add(1, Ordering::Relaxed));
         }
-        let logger = Box::new(telio::ffi_types::telio_logger_cb {
+        let logger = telio::ffi_types::telio_logger_cb {
             ctx: &call_count as *const AtomicUsize as *mut c_void,
             cb: test_telio_logger_fn,
-        });
-        log::set_boxed_logger(logger).unwrap();
-        tracing::subscriber::set_global_default(telio_utils::tracing::TracingToLogConverter)
-            .unwrap();
-        log::set_max_level(log::LevelFilter::Info);
+        };
+        let tracing_subscriber = TelioTracingSubscriber::new(logger, tracing::Level::INFO);
+        tracing::subscriber::set_global_default(tracing_subscriber).unwrap();
 
         tracing::info!("test message");
         assert_eq!(1, call_count.load(Ordering::Relaxed));