+++ changelog.md
@@ -7,6 +7,7 @@
 * LLT-4006: Restrict nat-lab containers capabilities
 * LLT-4586: Fix wireguard-go initialization crash
 * LLT-4634: Add wildcarded domains to DNS records
+* LLT-4225: Migrate Telio to use the tracing crate instead of log crate
 
 <br>
 
@@ -48,7 +49,6 @@
 * LLT-4502: Implement ICMP, UDP and TCP conntrack tracking for each peer separately
 * LLT-4508: Fetch windows dependencies from local gitlab registry
 * LLT-4552: Add way to flush events when stopping
-* LLT-4225: Migrate Telio to use the tracing crate instead of log crate
 * LLT-4583: Don't add IPv6 addresses to wireguard when IPv6 feature is disabled
 * LLT-4613: Fix boringtun performs spurious rekeying